refactor: add separate module for Atoms
