+++ whiplash/src/grammar/production/mod.rs
@@ -1,6 +1,7 @@
 pub mod rule;
 pub mod atom;
+pub mod atoms;
 
 pub use rule::Rule;
 pub use atom::Atom;
-pub use atom::Atoms;
+pub use atoms::Atoms;
